+++ src/main/kotlin/venus/riscv/insts/dsl/impls/extensions/FRTypeImplementation32.kt
@@ -3,12 +3,13 @@ package venus.riscv.insts.dsl.impls.extensions
 import venus.riscv.InstructionField
 import venus.riscv.MachineCode
 import venus.riscv.insts.dsl.impls.InstructionImplementation
+import venus.riscv.insts.floating.Decimal
 import venus.simulator.Simulator
 
 /**
  * Created by thaum on 8/6/2018.
  */
-class FRTypeImplementation32(private val eval: (Float, Float) -> Float) : InstructionImplementation {
+class FRTypeImplementation32(private val eval: (Decimal, Decimal) -> Decimal) : InstructionImplementation {
     override operator fun invoke(mcode: MachineCode, sim: Simulator) {
         val rs1 = mcode[InstructionField.RS1]
         val rs2 = mcode[InstructionField.RS2]
